A .bst file that limits output to the first three
This is a .bst that consists of the spbasic.bst created by Patrick W. Daly. It was under copyright from 1994-2004. I modified this code to only print out the first three names in a .bib author list if the list contains four or more authors. It will print out all the authors for an input list of three or less.